PACT-381 - EPR - Rendering Edit Set from model
The Edit Set page now renders a table according to GDS components that takes a EntrySet model object and creates a row in the table iteratively from a sequence of EntrySetEntry objects.

Also the heading of the page is now created from a parametrized message along with the Entry Set name.