PACT-442 | EPR | Creation of session token after login
Landing page on HomeController now checks the session token and,
if the user is authenticated, it's redirected to the editset page
(the current internal landing page), otherwise it gets redirected to
login.

After verifying credentials being valid, a new session token is created.
Then the user is redirected into the editset page with a session. This
page now sits on a private area of the website, only accesible with
correct usernames and passwords.