PACT-442 | EPR | Creation of session token after login
A new Secured trait has been created to encapsulate the logic that
checks authentication using the session token. Controllers that
require certain routes to be secured can now extend this trait.

Existing controller unit test have been updated to reflect the fact
that a valid session token is required to access private routes.

New unit tests have been created that validate the redirections that
happen when an secured route is accessed with an unauthenticated request.

Some unit controller and view tests that used localized messages have
been patched to make them work with the default locale of the machine.
Those tests were failing on machines that did not have "en" as a locale.