PACT-445 | EPR | Loading data from case classes
There is now a package object that contains the data for the edit set
records that gets loaded into the forms. This data is indexed using the
 OCI code that will be used to map incoming request and populate the
 form accordingly.

Also the data displayed on the discard and save pages is loaded from
the data map and links to return to previous page are fixed.